% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_obs.R
\name{load_obs}
\alias{load_obs}
\title{Load observations}
\usage{
load_obs(obs_daily_path, obs_sum_path, project_path, dailyVar, sumVar)
}
\arguments{
\item{obs_daily_path}{Path to the file of daily observations}

\item{obs_sum_path}{Path to the file of summary observations}

\item{project_path}{Path to the directory of input files}

\item{dailyVar}{Vector of the names of variables to estimate (variables from daily file)}

\item{sumVar}{Vector of the names of variables to estimate (variables from summary file)}
}
\value{
List of observed values to use for parameter estimation
A \verb{named list} (names = situations names) of data.frame containing
one column named Date with the dates (Date or POSIXct format) of the different observations
and one column per observed variables with either the measured values or NA, if
the variable is not observed at the given date
}
\description{
Load observations
}
