% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SQ.R
\name{run_SQ}
\alias{run_SQ}
\title{Build and call the command line}
\usage{
run_SQ(param_values, model_options, sim_output_path, situation)
}
\arguments{
\item{param_values}{A named vector that contains the value(s) and name(s)
of the parameters to force for each situation to simulate.}

\item{model_options}{Options for the model wrapper}

\item{sim_output_path}{Path to where we save the simulations}

\item{situation}{Situation to simulate}
}
\description{
Build and call the command line
}
