% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_console_data.R
\name{set_console_data}
\alias{set_console_data}
\title{Set console_data used to build the command line}
\usage{
set_console_data(run2simulate, project_path)
}
\arguments{
\item{run2simulate}{Run name to simulate}

\item{project_path}{Path to the directory of input files}
}
\value{
a list containing :
\code{run_names} list of the runs to simulate
\code{sit2simulate}list of situations to simulate
\code{Var} list of items for varietal parameters
\code{NonVar} list of items for non-varietal parameters
\code{Man} list of items for farming practices parameters
\code{Soil} list of items for soil parameters
\code{Site} list of items for site parameters
}
\description{
Set console_data used to build the command line
}
