% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_model_options.R
\name{set_model_options}
\alias{set_model_options}
\title{Set options for the model wrapper}
\usage{
set_model_options(
  cores_nb,
  time_display,
  project_path,
  output_path,
  console_path,
  saves_daily_output,
  print_console,
  Var_simule,
  Var_names,
  Param_list,
  console_data
)
}
\arguments{
\item{cores_nb}{Number of cores to simulate on}

\item{time_display}{Flag (TRUE/FALSE) for display of calculation time}

\item{project_path}{Path to the directory of input files of SQ}

\item{output_path}{Path where to save results}

\item{console_path}{Path to the console}

\item{saves_daily_output}{Boolean to save daily output (if set to "true")}

\item{print_console}{Boolean to print information on the simulation progression on the console (if set to "false")}

\item{Var_simule}{Names of variables to simulate used in the output files of SQ}

\item{Var_names}{Names of variables to simulate used in the observations files}

\item{Param_list}{A data frame containing the names of parameters to estimate and their types}

\item{console_data}{Set of lists needed to build the command line}
}
\value{
List containing options needed to run the model
}
\description{
Set options for the model wrapper
}
