% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_optim_options.R
\name{set_optim_options}
\alias{set_optim_options}
\title{Set options for the estimation method}
\usage{
set_optim_options(
  optim_method = "dreamzs",
  path_results = getwd(),
  ranseed = NULL,
  iterations = NULL,
  nCR = NULL,
  updateInterval = NULL,
  gamma = NULL,
  Eps = NULL,
  e = NULL,
  pCRupdate = NULL,
  burnin = NULL,
  thin = NULL,
  adaptation = NULL,
  DEpairs = NULL,
  ZupdateFrequency = NULL,
  pSnooker = NULL,
  Z = NULL,
  startValue = NULL,
  consoleUpdates = NULL,
  message = NULL,
  nb_rep = 5,
  xtol_rel = 1e-04,
  maxeval = 500
)
}
\arguments{
\item{optim_method}{Estimation method (dreamzs or simplex)}

\item{path_results}{Path where to store results and graphs}

\item{ranseed}{Seed for random numbers}

\item{iterations}{Total number of iterations}

\item{nCR}{Parameter determining the number of cross-over proposals. If nCR = 1 all parameters are updated jointly.}

\item{updateInterval}{Determining the intervall for the pCR (crossover probabilities)}

\item{gamma}{Kurtosis parameter Bayesian Inference Scheme.}

\item{Eps}{Ergodicity term}

\item{e}{Ergodicity term}

\item{pCRupdate}{Update of crossover probabilities}

\item{burnin}{Number of iterations treated as burn-in}

\item{thin}{Thin thinning parameter. Determines the interval in which values are recorded.}

\item{adaptation}{Number or percentage of samples that are used for the adaptation in DREAM}

\item{DEpairs}{Number of pairs used to generate proposal}

\item{ZupdateFrequency}{Frequency to update Z matrix}

\item{pSnooker}{Probability of snooker update}

\item{Z}{Starting matrix for Z}

\item{startValue}{Number of markov chains}

\item{consoleUpdates}{Intervall in which the sampling progress is printed to the console}

\item{message}{Logical determines whether the sampler's progress should be printed}

\item{nb_rep}{Number of minimization repetitions}

\item{xtol_rel}{Stop on small optimization step}

\item{maxeval}{Stop on this many function evaluations}
}
\value{
List of options for the estimation method
}
\description{
Set options for the estimation method
}
