% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_param_list.R
\name{set_param_list}
\alias{set_param_list}
\title{Set the data frame which associate each parameter to its type extracted from the project files}
\usage{
set_param_list(param_info, project_path)
}
\arguments{
\item{param_info}{Information on the parameters to estimate.
#' Either
a list containing:
\itemize{
\item (named) vectors of upper and lower bounds (\code{ub} and \code{lb}) (-Inf and Inf can be used),
\item \code{init_values}, A data.frame containing initial
values to test for the parameters (optional, if not provided, or if less values
than number of repetitions of the minimization are provided, the, or part
of the, initial values will be randomly generated using LHS sampling within
parameter bounds).
}

or
a named list containing for each parameter the list of situations per group
(\code{sit_list}), the vector of upper and lower bounds (one value per group)
(\code{ub} and \code{lb}) and the list of initial values per group
\code{init_values} (data.frame, one column per group, optional).
(see \href{https://sticsrpacks.github.io/CroptimizR/articles/Parameter_estimation_Specific_and_Varietal.html}{here}
for an example)}

\item{project_path}{Path to the project directory of input files}
}
\value{
A data frame containing
two columns :  \code{name} with the names of parameters and \code{type} with the types of each parameter
either Var, NonVar, Soil, Site or Man
}
\description{
Set the data frame which associate each parameter to its type extracted from the project files
}
