% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_runs_list.R
\name{set_runs_list}
\alias{set_runs_list}
\title{Set the list of items associated to a given run name (situations,Var,NonVar,Site,Soil,Man)}
\usage{
set_runs_list(run2simulate, project_path)
}
\arguments{
\item{run2simulate}{Run name to simulate}

\item{project_path}{Path to the directory of input files}
}
\value{
A \verb{named list} (names = runs names) of data frame containing the columns :
\code{situation} the associated situations
\code{Var}  items for varietal parameters
\code{NonVar} items for non-varietal parameters
\code{Soil} items for soil parameters
\code{Site} items for site parameters
\code{Man} items for managnement parameters
}
\description{
Set the list of items associated to a given run name (situations,Var,NonVar,Site,Soil,Man)
}
