% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_read.R
\name{sim_read}
\alias{sim_read}
\title{Read the simulations (output files of SQ)}
\usage{
sim_read(sim_output_path, model_options)
}
\arguments{
\item{sim_output_path}{Path to the output file}

\item{model_options}{Options for the model wrapper}
}
\value{
A named list (names = situations names )
Each element of the list should contains a data.frame with the results obtained for all simulated variables and dates for the given situation.
The data.frames must have one column called \code{Date} containing the simulations dates, in Date or POSIXct format
}
\description{
Read the simulations (output files of SQ)
}
