% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQ_optim.R
\name{SQ_optim}
\alias{SQ_optim}
\title{Main function for parameter estimation}
\usage{
SQ_optim(
  crit_function,
  optim_method,
  obs_daily_path,
  obs_sum_path,
  dailyVar,
  sumVar,
  obs_sim_path,
  project_path,
  output_path,
  console_path,
  param_info,
  optim_options,
  run2simulate,
  cores_nb,
  time_display,
  saves_daily_output,
  print_console
)
}
\arguments{
\item{crit_function}{Criterium to optimise depends on the chosen method :
Simplex : "crit_log_cwss", "crit_log_cwss_corr", "crit_ols"
Dreamzs : "likelihood_log_ciidn" , "likelihood_log_ciidn_corr"}

\item{optim_method}{Estimation method (Dreamzs, Simplex)}

\item{obs_daily_path}{Path to daily observations}

\item{obs_sum_path}{Path to summary observations}

\item{dailyVar}{Vector of variables names to estimate (variables from daily file)}

\item{sumVar}{Vector of the variables names to estimate (variables from summary file)}

\item{obs_sim_path}{Path to the data frame which matches between the names of the simulated and observed variables}

\item{project_path}{Path to the directory of input files}

\item{output_path}{Path where to save results}

\item{console_path}{Path to the console}

\item{param_info}{Information on the parameters to estimate.
#' Either
a list containing:
\itemize{
\item (named) vectors of upper and lower bounds (\code{ub} and \code{lb}) (-Inf and Inf can be used),
\item \code{init_values}, A data.frame containing initial
values to test for the parameters (optional, if not provided, or if less values
than number of repetitions of the minimization are provided, the, or part
of the, initial values will be randomly generated using LHS sampling within
parameter bounds).
}

or
a named list containing for each parameter the list of situations per group
(\code{sit_list}), the vector of upper and lower bounds (one value per group)
(\code{ub} and \code{lb}) and the list of initial values per group
\code{init_values} (data.frame, one column per group, optional).
(see \href{https://sticsrpacks.github.io/CroptimizR/articles/Parameter_estimation_Specific_and_Varietal.html}{here}
for an example)}

\item{optim_options}{List of options of the parameter estimation method, containing:
\itemize{
\item \code{path_results} The path where to store the results (optional, default=getwd())
\item specific options depending on the method used. Click on the links to see examples with the \href{https://sticsrpacks.github.io/CroptimizR/articles/Parameter_estimation_simple_case.html}{simplex}
and \href{https://sticsrpacks.github.io/CroptimizR/articles/Parameter_estimation_DREAM.html}{DreamZS} methods.
}}

\item{run2simulate}{Run name to simulate}

\item{cores_nb}{Number of cores to simulate on}

\item{time_display}{Flag (TRUE/FALSE) for display of calculation time}

\item{saves_daily_output}{Boolean to save daily output (if set to "true")}

\item{print_console}{Boolean to print information on the simulation progression on the console (if set to "false")}
}
\value{
prints, graphs and a list containing the results of the parameter estimation,
All results are saved in the folder \code{optim_options$path_results}.
}
\description{
Main function for parameter estimation
}
