% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQ_wrapper.R
\name{SQ_wrapper}
\alias{SQ_wrapper}
\title{SiruisQuality wrapper for CroptimizR}
\usage{
SQ_wrapper(
  param_values = NULL,
  sit_names = NULL,
  sit_var_dates_mask = NULL,
  model_options,
  ...
)
}
\arguments{
\item{param_values}{(optional) a named vector that contains the value(s) and name(s)
of the parameters to force for each situation to simulate. If not provided (or if is
NULL), the simulations will be performed using default values of the parameters
(e.g. as read in the model input files).}

\item{sit_names}{Vector of situations names for which results must be returned}

\item{sit_var_dates_mask}{mask to filter situations}

\item{model_options}{options for the model wrapper}

\item{...}{other options}
}
\value{
A list containing:
\code{sim_list}: a \verb{named list} (names = situations names) of data.frames (or tibbles) of
\code{error}: an error code indicating if at least one simulation ended with an error.
}
\description{
This function runs SiruisQuality on a set of situations (i.e. environments)
using the values of the parameters defined in the param_values argument. It returns
the values of the simulated outputs.
}
